+++ src/storage/core.rs
@@ -1,4 +1,8 @@
+use crate::error::ValidationErrorKind;
+
 use super::prelude::*;
+use futures::stream::FuturesOrdered;
+use tokio::fs::{create_dir, create_dir_all};
 
 const BLOB_FILE_EXTENSION: &str = "blob";
 const LOCK_FILE: &str = "pearl.lock";
@@ -9,7 +13,7 @@ const O_EXCL: i32 = 128;
 ///
 /// This type is clonable, cloning it will only create a new reference,
 /// not a new storage.
-/// Storage has a type parameter K.
+/// Storage has a type kindeter K.
 /// To perform read/write operations K must implement [`Key`] trait.
 ///
 /// # Examples
@@ -105,19 +109,22 @@ impl<K: Key> Storage<K> {
     /// storage creates it, otherwise tries to init existing storage.
     /// # Errors
     /// Returns error in case of failures with IO operations or
-    /// if some of the required params are missed.
+    /// if some of the required kinds are missed.
     ///
     /// [`init()`]: struct.Storage.html#method.init
     pub async fn init(&mut self) -> Result<()> {
         // @TODO implement work dir validation
-        self.prepare_work_dir().await?;
-        let cont_res = work_dir_content(
-            self.inner
-                .config
-                .work_dir()
-                .ok_or_else(|| Error::from(ErrorKind::Uninitialized))?,
-        )
-        .await;
+        self.prepare_work_dir()
+            .await
+            .context("failed to prepare work dir")?;
+        let wd = self
+            .inner
+            .config
+            .work_dir()
+            .ok_or_else(|| Error::from(ErrorKind::Uninitialized))?;
+        let cont_res = work_dir_content(wd)
+            .await
+            .with_context(|| format!("failed to read work dir content: {}", wd.display()));
         trace!("work dir content loaded");
         if let Some(files) = cont_res? {
             trace!("storage init from existing files");
@@ -288,7 +295,7 @@ impl<K: Key> Storage<K> {
             .active_blob
             .as_ref()
             .ok_or_else(Error::active_blob_not_set)?
-            .read_any(key, meta)
+            .read_any(key, meta, true)
             .await;
         debug!("storage read with optional meta from active blob finished");
         match active_blob_read_res {
@@ -303,16 +310,52 @@ impl<K: Key> Storage<K> {
         }
     }
 
-    async fn get_any_data(safe: &Safe, key: &[u8], meta: Option<&Meta>) -> Result<Vec<u8>> {
+    async fn get_data_last(safe: &Safe, key: &[u8], meta: Option<&Meta>) -> Result<Vec<u8>> {
         let blobs = safe.blobs.read().await;
-        let stream: FuturesUnordered<_> =
-            blobs.iter().map(|blob| blob.read_any(key, meta)).collect();
+        let blobs_stream: FuturesOrdered<_> = blobs
+            .iter()
+            .enumerate()
+            .map(|(i, blob)| async move {
+                if blob.check_filters_non_blocking(key).await {
+                    Some(i)
+                } else {
+                    None
+                }
+            })
+            .collect();
+        let possible_blobs: Vec<_> = blobs_stream.filter_map(|e| e).collect().await;
+        debug!(
+            "len of possible blobs: {} (start len: {})",
+            possible_blobs.len(),
+            blobs.len()
+        );
+        let stream: FuturesOrdered<_> = possible_blobs
+            .iter()
+            .rev()
+            .map(|i| blobs[*i].read_any(key, meta, false))
+            .collect();
+        debug!("read with optional meta {} closed blobs", stream.len());
+        let mut task = stream.skip_while(Result::is_err);
+        task.next().await.ok_or_else(Error::not_found)?
+    }
+
+    #[allow(dead_code)]
+    async fn get_data_any(safe: &Safe, key: &[u8], meta: Option<&Meta>) -> Result<Vec<u8>> {
+        let blobs = safe.blobs.read().await;
+        let stream: FuturesUnordered<_> = blobs
+            .iter()
+            .map(|blob| blob.read_any(key, meta, true))
+            .collect();
         debug!("read with optional meta {} closed blobs", stream.len());
         let mut task = stream.skip_while(Result::is_err);
         task.next()
             .await
             .ok_or_else(Error::not_found)?
             .with_context(|| "no results in closed blobs")
+    }
+
+    async fn get_any_data(safe: &Safe, key: &[u8], meta: Option<&Meta>) -> Result<Vec<u8>> {
+        Self::get_data_last(safe, key, meta).await
     }
 
     /// Stop blob updater and release lock file
@@ -389,7 +432,7 @@ impl<K: Key> Storage<K> {
             debug!("work dir exists: {}", path.display());
         } else if self.inner.config.create_work_dir() {
             debug!("creating work dir recursively: {}", path.display());
-            std::fs::create_dir_all(path)?;
+            create_dir_all(path).await?;
         } else {
             error!("work dir path not found: {}", path.display());
             return Err(Error::work_dir_unavailable(
@@ -399,7 +442,9 @@ impl<K: Key> Storage<K> {
             ))
             .with_context(|| "failed to prepare work dir");
         }
-        self.try_lock_dir(path).await
+        self.try_lock_dir(path)
+            .await
+            .with_context(|| format!("failed to lock work dir: {}", path.display()))
     }
 
     async fn try_lock_dir<'a>(&'a self, path: &'a Path) -> Result<()> {
@@ -423,9 +468,8 @@ impl<K: Key> Storage<K> {
 
     async fn init_new(&mut self) -> Result<()> {
         let next = self.inner.next_blob_name()?;
-        let config = self.filter_config();
         let mut safe = self.inner.safe.write().await;
-        let blob = Blob::open_new(next, self.inner.ioring.clone(), config)
+        let blob = Blob::open_new(next, self.inner.ioring.clone(), self.inner.config.index())
             .await?
             .boxed();
         safe.active_blob = Some(blob);
@@ -492,7 +536,7 @@ impl<K: Key> Storage<K> {
             .map(|file| async {
                 let sem = disk_access_sem.clone();
                 let _sem = sem.acquire().await.expect("sem is closed");
-                Blob::from_file(file.clone(), ioring.clone(), config.filter(), K::LEN)
+                Blob::from_file(file.clone(), ioring.clone(), config.index(), K::LEN)
                     .await
                     .map_err(|e| (e, file))
             })
@@ -503,12 +547,20 @@ impl<K: Key> Storage<K> {
             match blob_res {
                 Ok(blob) => blobs.push(blob),
                 Err((e, file)) => {
-                    let msg = format!(
-                        "Failed to read existing blob!\nPath: {:?};\nReason: {:?}.",
-                        file, e
-                    );
+                    let msg = format!("Failed to read existing blob: {}", file.display());
                     if config.ignore_corrupted() {
-                        error!("{}", msg);
+                        error!("{}, cause: {:#}", msg, e);
+                    } else if Self::should_save_corrupted_blob(&e) {
+                        error!(
+                            "save corrupted blob '{}' to directory '{}'",
+                            file.display(),
+                            config.corrupted_dir_name()
+                        );
+                        Self::save_corrupted_blob(&file, config.corrupted_dir_name())
+                            .await
+                            .with_context(|| {
+                                anyhow!(format!("failed to save corrupted blob {:?}", file))
+                            })?;
                     } else {
                         return Err(e.context(msg));
                     }
@@ -516,6 +568,63 @@ impl<K: Key> Storage<K> {
             }
         }
         Ok(blobs)
+    }
+
+    fn should_save_corrupted_blob(error: &anyhow::Error) -> bool {
+        debug!("decide wether to save corrupted blobs: {:#}", error);
+        if let Some(error) = error.downcast_ref::<Error>() {
+            return match error.kind() {
+                ErrorKind::Bincode(_) => true,
+                ErrorKind::Validation { kind, cause: _ } => {
+                    !matches!(kind, ValidationErrorKind::BlobVersion)
+                }
+                _ => false,
+            };
+        }
+        false
+    }
+
+    async fn save_corrupted_blob(path: &Path, corrupted_dir_name: &str) -> Result<()> {
+        let parent = path
+            .parent()
+            .ok_or_else(|| anyhow!("[{}] blob path don't have parent directory", path.display()))?;
+        let file_name = path
+            .file_name()
+            .ok_or_else(|| anyhow!("[{}] blob path don't have file name", path.display()))?
+            .to_os_string();
+        let corrupted_dir_path = parent.join(corrupted_dir_name);
+        let corrupted_path = corrupted_dir_path.join(file_name);
+        if corrupted_dir_path.exists() {
+            debug!("{} dir exists", path.display());
+        } else {
+            debug!("creating dir for corrupted files: {}", path.display());
+            create_dir(corrupted_dir_path).await.with_context(|| {
+                format!(
+                    "failed to create dir for corrupted files: {}",
+                    path.display()
+                )
+            })?;
+        }
+        tokio::fs::rename(&path, &corrupted_path)
+            .await
+            .with_context(|| {
+                anyhow!(format!(
+                    "failed to move file {:?} to {:?}",
+                    path, corrupted_path
+                ))
+            })?;
+        Self::remove_index_by_blob_path(path).await?;
+        Ok(())
+    }
+
+    async fn remove_index_by_blob_path(path: &Path) -> Result<()> {
+        let index_path = path.with_extension(blob::BLOB_INDEX_FILE_EXTENSION);
+        if index_path.exists() {
+            tokio::fs::remove_file(&index_path)
+                .await
+                .with_context(|| anyhow!(format!("failed to remove file {:?}", index_path)))?;
+        }
+        Ok(())
     }
 
     /// `contains` is used to check whether a key is in storage.
@@ -545,27 +654,25 @@ impl<K: Key> Storage<K> {
         Ok(false)
     }
 
-    /// `check_bloom` is used to check whether a key is in storage.
-    /// If bloom filter opt out, returns `None`.
-    /// Uses bloom filter under the hood, so false positive results are possible,
-    /// but false negatives are not.
-    /// In other words, `check_bloom` returns either "possibly in storage" or "definitely not".
-    pub async fn check_bloom(&self, key: impl AsRef<K>) -> Option<bool> {
+    /// `check_filters` is used to check whether a key is in storage.
+    /// Range (min-max test) and bloom filters are used.
+    /// If bloom filter opt out and range filter passes, returns `None`.
+    /// False positive results are possible, but false negatives are not.
+    /// In other words, `check_filters` returns either "possibly in storage" or "definitely not".
+    pub async fn check_filters(&self, key: impl AsRef<K>) -> Option<bool> {
         let key = key.as_ref();
         trace!("[{:?}] check in blobs bloom filter", &key.to_vec());
         let inner = self.inner.safe.read().await;
         let in_active = inner
             .active_blob
             .as_ref()
-            .map_or(Some(false), |active_blob| {
-                active_blob.check_bloom(key.as_ref())
-            })?;
+            .map_or(false, |active_blob| active_blob.check_filters(key.as_ref()));
         let in_closed = inner
             .blobs
             .read()
             .await
             .iter()
-            .any(|blob| blob.check_bloom(key.as_ref()) == Some(true));
+            .any(|blob| blob.check_filters(key.as_ref()) == true);
         Some(in_active || in_closed)
     }
 
@@ -600,10 +707,6 @@ impl<K: Key> Storage<K> {
     /// Or if there are any problems with syncronization.
     pub async fn close_active_blob(&self) {
         self.observer.close_active_blob().await
-    }
-
-    fn filter_config(&self) -> Option<BloomConfig> {
-        self.inner.config.filter()
     }
 
     fn launch_observer(&mut self) {