+++ tests/tests.rs
@@ -139,7 +139,7 @@ async fn test_multithread_read_write() -> Result<(), String> {
         .collect();
     let handles = handles.try_collect::<Vec<_>>().await.unwrap();
     let index = path.join("test.0.index");
-    sleep(Duration::from_millis(32)).await;
+    sleep(Duration::from_millis(64)).await;
     assert!(index.exists());
     assert_eq!(handles.len(), threads);
     let keys = indexes
@@ -238,6 +238,33 @@ async fn test_work_dir_lock() {
     dbg!(&res_two);
     assert!(res_two.is_err());
     common::clean(storage, path)
+        .map(|res| res.expect("clean failed"))
+        .await;
+    warn!("elapsed: {:.3}", now.elapsed().as_secs_f64());
+}
+
+#[tokio::test]
+async fn test_corrupted_index_regeneration() {
+    let now = Instant::now();
+    let path = common::init("corrupted_index");
+    let storage = common::create_test_storage(&path, 70_000).await.unwrap();
+    let records = common::generate_records(100, 10_000);
+    for (i, data) in &records {
+        write_one(&storage, *i, data, None).await.unwrap();
+        sleep(Duration::from_millis(10)).await;
+    }
+    storage.close().await.unwrap();
+    let index_file_path = path.join("test.0.index");
+    assert!(index_file_path.exists());
+
+    common::corrupt_file(index_file_path, common::CorruptionType::ZeroedAtBegin(1024))
+        .expect("index corruption failed");
+
+    let new_storage = common::create_test_storage(&path, 1_000_000)
+        .await
+        .expect("storage should be loaded successfully");
+
+    common::clean(new_storage, path)
         .map(|res| res.expect("clean failed"))
         .await;
     warn!("elapsed: {:.3}", now.elapsed().as_secs_f64());
@@ -481,11 +508,11 @@ async fn test_check_bloom_filter_single() {
         trace!("key: {}, pos: {:?}, negative: {:?}", i, pos_key, neg_key);
         let key = KeyTest::new(i);
         storage.write(&key, data.to_vec()).await.unwrap();
-        assert_eq!(storage.check_bloom(key).await, Some(true));
+        assert_eq!(storage.check_filters(key).await, Some(true));
         let data = b"other_random_data";
         storage.write(&pos_key, data.to_vec()).await.unwrap();
-        assert_eq!(storage.check_bloom(pos_key).await, Some(true));
-        assert_eq!(storage.check_bloom(neg_key).await, Some(false));
+        assert_eq!(storage.check_filters(pos_key).await, Some(true));
+        assert_eq!(storage.check_filters(neg_key).await, Some(false));
     }
     common::clean(storage, path).await.expect("clean failed");
     warn!("elapsed: {:.3}", now.elapsed().as_secs_f64());
@@ -504,10 +531,10 @@ async fn test_check_bloom_filter_multiple() {
         trace!("blobs count: {}", storage.blobs_count().await);
     }
     for i in 1..800 {
-        assert_eq!(storage.check_bloom(KeyTest::new(i)).await, Some(true));
+        assert_eq!(storage.check_filters(KeyTest::new(i)).await, Some(true));
     }
     for i in 800..1600 {
-        assert_eq!(storage.check_bloom(KeyTest::new(i)).await, Some(false));
+        assert_eq!(storage.check_filters(KeyTest::new(i)).await, Some(false));
     }
     common::clean(storage, path).await.expect("clean failed");
     warn!("elapsed: {:.3}", now.elapsed().as_secs_f64());
@@ -540,13 +567,13 @@ async fn test_check_bloom_filter_init_from_existing() {
     let storage = common::create_test_storage(&path, 100).await.unwrap();
     debug!("check check_bloom");
     for i in 1..base {
-        assert_eq!(storage.check_bloom(KeyTest::new(i)).await, Some(true));
+        assert_eq!(storage.check_filters(KeyTest::new(i)).await, Some(true));
     }
     info!("check certainly missed keys");
     let mut false_positive_counter = 0usize;
     for i in base..base * 2 {
         trace!("check key: {}", i);
-        if storage.check_bloom(KeyTest::new(i)).await == Some(true) {
+        if storage.check_filters(KeyTest::new(i)).await == Some(true) {
             false_positive_counter += 1;
         }
     }
@@ -588,13 +615,13 @@ async fn test_check_bloom_filter_generated() {
     debug!("check check_bloom");
     for i in 1..base {
         trace!("check key: {}", i);
-        assert_eq!(storage.check_bloom(KeyTest::new(i)).await, Some(true));
+        assert_eq!(storage.check_filters(KeyTest::new(i)).await, Some(true));
     }
     info!("check certainly missed keys");
     let mut false_positive_counter = 0usize;
     for i in base..base * 2 {
         trace!("check key: {}", i);
-        if storage.check_bloom(KeyTest::new(i)).await == Some(true) {
+        if storage.check_filters(KeyTest::new(i)).await == Some(true) {
             false_positive_counter += 1;
         }
     }
@@ -810,5 +837,59 @@ async fn test_memory_index() {
         KEY_AND_DATA_SIZE * 3 + RECORD_HEADER_SIZE * 3
     ); // 3 keys, 3 records in active blob (3 allocated)
     assert!(path.join("test.1.blob").exists());
+    common::clean(storage, path).await.unwrap();
+}
+
+#[tokio::test]
+async fn test_blob_header_validation() {
+    use pearl::error::{AsPearlError, ValidationErrorKind};
+    use std::os::unix::fs::FileExt;
+
+    let path = common::init("blob_header_validation");
+    let storage = common::create_test_storage(&path, 10_000).await.unwrap();
+    let data = vec![1, 1, 2, 3, 5, 8];
+    write_one(&storage, 42, &data, None).await.unwrap();
+    storage.close().await.expect("storage close failed");
+    let blob_path = std::env::temp_dir().join(&path).join("test.0.blob");
+    info!("path: {}", blob_path.display());
+    let file = fs::OpenOptions::new()
+        .write(true)
+        .create(false)
+        .open(&blob_path)
+        .expect("failed to open file");
+    let buf = bincode::serialize(&0_u32).expect("failed to serialize u32");
+    file.write_at(&buf, 8)
+        .expect("failed to overwrite blob version");
+
+    let builder = Builder::new()
+        .work_dir(&path)
+        .blob_file_name_prefix("test")
+        .max_blob_size(10_000)
+        .max_data_in_blob(100_000)
+        .set_filter_config(Default::default())
+        .allow_duplicates();
+    let builder = if let Ok(ioring) = rio::new() {
+        builder.enable_aio(ioring)
+    } else {
+        println!("current OS doesn't support AIO");
+        builder
+    };
+    let mut storage: Storage<KeyTest> = builder.build().unwrap();
+    let err = storage
+        .init()
+        .await
+        .expect_err("storage initialized with invalid blob header");
+    info!("{:#}", err);
+    let pearl_err = err
+        .as_pearl_error()
+        .expect("result doesn't contains pearl error");
+    let is_correct = matches!(
+        pearl_err.kind(),
+        pearl::ErrorKind::Validation {
+            kind: ValidationErrorKind::BlobVersion,
+            cause: _
+        }
+    );
+    assert!(is_correct);
     common::clean(storage, path).await.unwrap();
 }